server/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron_descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-configurations==2.5
django-cors-headers==4.3.1
django-filter==23.5
django-timezone-field==6.1.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
kombu==5.3.5
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.17
python-crontab==3.0.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
whitenoise==6.6.0
